Tryout of fetching the providerVerticals in the SearchStore.
I am not sure whether this is the correct approach, since we really
want this fetch to only happen once when the frontend server is started,
while doing it in the SearchStore triggers the fetch on each page load.
However, react needs to be aware that an update of state happens after
the request returns, so I'm not entirely sure what a better approach is.