Add autohide option
If autohide is set to 'individual', results bookmarked or excluded
by the user are hidden automatically. If set to 'shared' results
bookmarked or excluded by anyone in the group are hidden automatically.

This gives us a way to trade off efficiency of the distribution
of labour vs having a consitent page shown to the user.