Fix search query resetting on SearchStore update
When the searchstore updates, we compared whether the SearchState
has changed to update the query when a user clicks an item in the
query history. However, the query field of the SearchState was also
used as state of the currently entered query after a user has edited
the query. This caused the query to reset each time the SearchStore
updates (which for example happens when another user in the group
issues a new query). Fixed by using a seperate field for the current
query state.