Fix real-time update of annotations and ratings
Since we no longer re-fresh the entire results list when a page
metadata update is triggered, real-time updating of annotations
and ratings was broken. By merging the data from the annotation
and ratings stores into the result list on a page metadata update,
this issue is resolved. In the future we may want to remove the
annotation, rating, and bookmark stores altogether, and move to
a single state tree. This would remove the duplication of state
between the result list in the SearchStore and the other stores.