Fix clicking bookmarks by fetching missing Annotations and Ratings
Clicking bookmarks that are added by another user was broken
when annotations or ratings were activated. This issue did not
occur before since we did not use ratings and annotations in the
algorithmic mediation experiment, and is caused because we had
to change the way metadata is updated when algorithmic mediation
was added.