Fix CORS issue by setting headers to allow cross-origin requests
Firefox was blocking requests to the backend server, because it
is running on another domain than the frontend server. Setting
headers to allow cross-origin requests fixes this for development.
In a production application CORS should be limited to specific
domains, but the development server should not be used in a
production setting anyways.